fix: Refactor to prevent cyclic dependencies
